requirements.txt
flask==0.10.1
flask-SQLAlchemy==1.0
markdown==2.5.1
pyembed-markdown==1.1.0
requests==2.4.3
beautifulsoup4==4.3.2
pyyaml==3.11
pillow==5.0.0
